stats

+++ shared/docker.go
@@ -2,7 +2,7 @@ package shared
 
 // ContainerStats struct
 type ContainerStats struct {
-	RAMUsage float64 `json:"ram_usage"`
-	MaxRAM   float64 `json:"max_ram"`
+	RAMUsage int64   `json:"ram_usage"`
+	MaxRAM   int64   `json:"max_ram"`
 	CPU      float64 `json:"cpu_usage"`
 }
